Use /craft_XXXXXX name for working folder
This change makes the working folder to change between projects
and to have a non-trivial name, to ensure that it isn't used
accidentally in a snapcraft.yaml file instead of the CRAFT_XXXXX
environment variables. It also notifies the user if the working
folder name appears inside the snapcraft.yaml file.